Remove listener from throttle manager
See GitHub Issue #10887
https://github.com/JMRI/JMRI/issues/10887
